Enable passing of AWS Session Token in Hazelcast AWS Config.
Our developers do not have permanent access and secret keys for connecting to AWS during development, which normally requires the use of AWS SDK Credentials Provider with STS Credentials. However, since Hazelcast AWS does not use the AWS SDK, this cannot be used to solve the issue. To fix this, the session token needs to be passed through as part of the Hazelcast AWS Config. This change enables the passing of the AWS Session Token, which now allows developers to use Hazelcast AWS with STS Credentials.

Implementation details: Added support for session token in Hazelcast AWS Config.